chore(KitDraggable): better feedback when not moving far away enough
When the user starts dragging, the ghost is shown when the position of the item would change when dropped. If the item is dragged over the items directly near the original position, there is no feedback about what will happen.

This commit changes the opacity of the original position of the dragged item to reflect the fact that it will move or not when dropped.