fix(KitModal): enhancing close-on-click-outside
If the user holds the mouse button down and moves it outside of the modal before releasing it, Chrome will report a click outsite the modal and Firefox will report a click on the originally clicked item.

Because of that, we fallback to monitoring mousedown and mouseup instead. mousedown registers the location and based on that mouseup closes or not the modal.