fix(KitSpotlight): avoid mask holes
With scaling (browser or OS), the positioning of the bottom part of the mask might get an of-by-one pixel, rendering a line between the bottom and the 2 middle masks.

Prior to this commit, the bottom mask was added with bottom: 0, and a height property.

This commit fixes this issue by using defining the correct top value instead of height.